New feature added: IConsoleBinder for parsing
In Console.as: Added a property for the binder and a function
bind(IConsoleBinder), as well as initializing the console with a default
echo binder. This changes the console listener so that commandes are
parsed by _consoleBinder.parseCommand().

In EchoBinder: Default binder that echoes every command back  the
console.

In IConsoleBinder: Interface that allows parsing.