Changed Binding feature
Now you bind functions directly into the Console. Each function is
binded to a string key and when executed, the binded function receives
an Array with the arguments. This array is the command introduced in the
text input splitted by SPACES.
You can use Wrapper functions to execute your system's methods from the
console.